+++ module/sheets/personnage-sheet.mjs
@@ -132,6 +132,7 @@ export class PersonnageSheet extends ActorSheet {
     ctx.adaptations = this.actor.items.filter(i => i.type === "adaptation");
     ctx.traumatismes = this.actor.items.filter(i => i.type === "traumatisme");
     ctx.historiques = this.actor.items.filter(i => i.type === "historique");
+    ctx.afflictions = this.actor.items.filter(i => i.type === "affliction");
 
     ctx.enrichedBio = await TextEditor.enrichHTML(sys.biographie ?? "", { async: true });
     ctx.enrichedNotes = await TextEditor.enrichHTML(sys.notes ?? "", { async: true });